requirements.txt
django-notifications-hq
asn1crypto
astroid
backports.functools-lru-cache==1.5
certifi
chardet
click
configparser
coreapi
coreapi-cli
coreschema
cryptography
cssselect2
defusedxml
Django
django-appconf
django-background-tasks
django-compat
django-cryptography
django-debug-toolbar
django-easy-pdf
django-fernet-fields
django-settings-export
django-sitetree
django-stronghold
djangorestframework
djangorestframework-jwt
dnspython
enum34
future
futures
gunicorn
html5lib
httplib2
idna
ipaddress
isort
itypes
Jinja2
jira
lazy-object-proxy
lxml
MarkupSafe
mccabe
oauthlib
openvas_lib
ordereddict
pbr
pdfrw
Pillow
psycopg2-binary
PyBurprestapi
pycparser==2.18
PyJWT
pylint
PyPDF2
Pyphen
python-nmap
python-owasp-zap-v2.4
python-taint
python-whois
pytz==2017.3
reportlab
requests
requests-oauthlib
requests-toolbelt
selenium
singledispatch
six
tinycss2
uritemplate
urllib3
webencodings
whitenoise
wrapt
xhtml2pdf
beautifulsoup4
django-import-export
django-s3-storage
jsonfield
waitress